Minor refactoring of the Solver

+++ include/simulator/simulator_utilities.h
@@ -27,8 +27,8 @@ Mid ConvolvePartialDiff(const Convolution& convolution,
                         const std::vector<std::vector<Mid>>& known_d_mids,
                         const std::vector<EmuAndMid>& input_mids,
                         const std::vector<std::vector<Mid>>& saved_mids,
-                        int mid_size,
-                        int diff_position);
+                        size_t mid_size,
+                        size_t diff_position);
 
 void FillDiffYMatrix(const std::vector<PositionOfSavedEmu>& Y_data,
                      const std::vector<std::vector<Mid>>& known_d_mids,
