+++ src/utilities/debug_utills/debug_prints.cpp
@@ -65,7 +65,7 @@ void PrintReactions(const std::vector<Reaction>& reactions) {
 
 void PrintReaction(const Reaction& reaction) {
     std::cout << reaction.id << " " << reaction.name << " ";
-    int left_counter = 0;
+    size_t left_counter = 0;
     for (Substrate sub : reaction.chemical_equation.left) {
         std::cout << sub.substrate_coefficient_ << " " << sub.name;
         ++left_counter;
@@ -74,7 +74,7 @@ void PrintReaction(const Reaction& reaction) {
         }
     }
     std::cout << " = ";
-    int right_counter = 0;
+    size_t right_counter = 0;
     for (Substrate sub : reaction.chemical_equation.right) {
         std::cout << sub.substrate_coefficient_ << " " << sub.name;
         ++right_counter;
@@ -83,7 +83,7 @@ void PrintReaction(const Reaction& reaction) {
         }
     }
     std::cout << ", ";
-    int left_atom_counter = 0;
+    size_t left_atom_counter = 0;
     for (Substrate sub : reaction.chemical_equation.left) {
         std::cout << sub.atom_coefficient_ << " " << sub.formula;
         ++left_atom_counter;
@@ -92,7 +92,7 @@ void PrintReaction(const Reaction& reaction) {
         }
     }
     std::cout << " = ";
-    int right_atom_counter = 0;
+    size_t right_atom_counter = 0;
     for (Substrate sub : reaction.chemical_equation.right) {
         std::cout << sub.atom_coefficient_ << " " << sub.formula;
         ++right_atom_counter;
